+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent unverified users from joining an organization or using a template [#591](https://github.com/PublicMapping/districtbuilder/pull/591)
 - Project districts geojson is now cached in the project table when updated [#594](https://github.com/PublicMapping/districtbuilder/pull/594)
 - Upgrade NestJS deps [#637](https://github.com/PublicMapping/districtbuilder/pull/637)
+- Refactor project evaluate mode views [#642](https://github.com/PublicMapping/districtbuilder/pull/642)
 
 ### Fixed